removed unnecessary numeric casts
git-svn-id: https://svn.code.sf.net/p/vassalengine/svn/VASSAL-src/trunk@9415 67b53d14-2c14-4ace-a08f-0dab2b34000c

+++ src/VASSAL/counters/FreeRotator.java
@@ -615,7 +615,7 @@ public class FreeRotator extends Decorator
       myAngle = p.x < origin.x ? -Math.PI/2.0 : Math.PI/2.0;
     }
     else {
-      myAngle = Math.atan((double)(p.x - origin.x) / (double)(origin.y - p.y));
+      myAngle = Math.atan((double)(p.x - origin.x) / (origin.y - p.y));
       if (origin.y < p.y) {
         myAngle += Math.PI;
       }
